agent/requirements.txt
claude-agent-sdk
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
python-dotenv==1.1.0
anyio==4.9.0
